Reorganized the code. Created a RegisteredUsers module
